+++ src/BootstrapAggregator.php
@@ -185,9 +185,7 @@ class BootstrapAggregator implements Estimator, Learner, Parallel, Persistable
             ));
         }
 
-        $aggregate = $this->backend->process();
-
-        $aggregate = array_map(null, ...$aggregate);
+        $aggregate = transpose($this->backend->process());
         
         switch ($this->type()) {
             case self::CLASSIFIER: