+++ src/functions.php
@@ -34,4 +34,27 @@ namespace Rubix\ML
     {
         return log(array_sum(array_map('exp', $values)));
     }
+
+    /**
+     * Transpose a 2-dimensional array i.e. rotate the data table.
+     *
+     * @param array $table
+     * @return array
+     */
+    function transpose(array $table) : array
+    {
+        if (count($table) > 1) {
+            return array_map(null, ...$table);
+        }
+
+        $n = count(reset($table));
+
+        $columns = [];
+
+        for ($i = 0; $i < $n; $i++) {
+            $columns[] = array_column($table, $i);
+        }
+
+        return $columns;
+    }
 }