+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Added deduplicate method to dataset object
     - Added unzip static factory for Labeled datasets from data table
     - Changed the order of t-SNE hyperparameters
+    - Added global transpose array helper function
 
 - 0.0.15-beta
     - Added Gaussian MLE anomaly detector